+++ src/components/modals/students/StudentCreate.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { Modal, Form, Input, Select } from "antd";
 import useStudents from "../../../hooks/useStudents";
-import { useGroups } from "../../../hooks/useGroups";
+import { useCourse } from "../../../hooks/useCourse";
 
 const StudentCreateModal = ({ open, onClose, branchId }) => {
   const [form] = Form.useForm();
   const { createStudentMutation } = useStudents(branchId);
-  const { groupsQuery } = useGroups(branchId, { enabled: !!branchId });
+  const { groupsQuery } = useCourse(branchId, { enabled: !!branchId });
 
   const handleOk = () => {
     form.validateFields().then((values) => {
